Renamed batch/batchErrors.ts to batch/errors.ts
