+++ src/core/batch/batch.ts
@@ -1,7 +1,7 @@
 import { ParquetReader, ParquetWriter, ParquetSchema } from "@dsnp/parquetjs";
 
 import { AnnouncementWithSignature, AnnouncementType, TypedAnnouncement } from "../announcements";
-import { MixedTypeBatchError, EmptyBatchError } from "./batchErrors";
+import { MixedTypeBatchError, EmptyBatchError } from "./errors";
 import { ConfigOpts, requireGetStore } from "../config";
 import { getSchemaFor, getBloomFilterOptionsFor, Schema, BloomFilterOptions } from "./parquetSchema";
 import { WriteStream } from "../store";
