+++ src/core/batch/errors.ts
